Rename package ocfweb.atool -> ocfweb.account
