+++ ocfweb/account/templates/chpass/success.html
